% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/major_locus_G_summary.R
\name{major_locus_G_summary}
\alias{major_locus_G_summary}
\title{Summary measures for the major locus, \eqn{G}, by relative type}
\usage{
major_locus_G_summary(mimsim.object, relative_type, include = "Ascertained")
}
\arguments{
\item{mimsim.object}{S4 class object resulting from a call to \code{\link{mimsim}}.}

\item{relative_type}{Relative type. One of: \code{"P"} (parents), \code{"S"} (siblings), \code{"O"} (offspring), \code{"GP"} (grandparents), \code{"Av"} (aunts and uncles by blood),\code{"C"} (cousins), \code{"MAv"} (aunts and uncles by marriage) or \code{"M"} (mates).}

\item{include}{character. 2 options: 1. \code{"Ascertained"} = include families containing an eligible and consenting individual (ascertained families). These families are stored in \code{mimsim.object@pedigree}. 2. \code{"Complete"} = include families containing an eligible individual. This combines families stored in \code{mimsim.object@pedigree} (consenting, and so ascertained, families) and \code{mimsim.object@pedigree_unascertained} (non-consenting, and so unascertained, families). Default is \code{include = "Ascertained"}.}
}
\value{
\code{major_locus_G_summary} returns an S4 class object. The number of slots depends on input variable \code{relative_type}. For \code{relative_type =} \code{"P"} (parents), \code{"GP"} (grandparents), \code{"Av"} (aunts and uncles by blood), \code{"MAv"} (aunts and uncles by marriage) or \code{"M"} (mates) there are 3 slots outputted. For \code{relative_type =} \code{"S"} (siblings), \code{"O"} (offspring) or \code{"C"} (cousins) there are 5 slots outputted.

Slots are:
\item{Description}{character. Brief overview of output.}
\item{G_summary1}{data table. Counts and proportions of observed major locus genotypes for \code{relative_type}.}
\item{G_summary2}{data table. Counts and proportions of observed major locus genotypes for \code{relative_type}, stratified by the major locus genotype of the proband.}
\item{G_summary3}{data table. Counts and proportions of observed major locus genotypes for \code{relative_type}, stratified by sibship size. Only available for \code{relative_type =} \code{"S"} (siblings), \code{"O"} (offspring) or \code{"C"} (cousins).}
\item{G_summary4}{data table. Counts and proportions of observed major locus genotypes for \code{relative_type}, stratified by the major locus genotype of the proband and sibship size. Only available for \code{relative_type =} \code{"S"} (siblings), \code{"O"} (offspring) or \code{"C"} (cousins).}

Column headings:
\itemize{
\item G; observed genotype of the relatives of type \code{relative_type},
\item Count; number of relatives of type \code{relative_type} with observed G,
\item Proportion; proportion of relatives of type \code{relative_type} with observed G,
\item Genotype_PB; the observed major locus genotypes of Probands (PB), and,
\item S; observed sibship size (when \code{relative_type} is \code{"S"}, \code{"O"} or \code{"C"}).
}
}
\description{
\code{major_locus_G_summary} produces summary tables of counts and proportions for the major locus genotypes of a user-specified relative type. Relative type is defined with respect to the proband, e.g. a sibling means 'a sibling of a proband'.
}
\details{
For a user-specified relative type, summary measures produced are counts and proportions for the major locus genotypes. Options for the relative type input (\code{relative_type}) are: \code{"P"} (parents), \code{"S"} (siblings), \code{"O"} (offspring), \code{"GP"} (grandparents), \code{"Av"} (aunts and uncles by blood),\code{"C"} (cousins), \code{"MAv"} (aunts and uncles by marriage) or \code{"M"} (mates). Only one relative type is specified.

The number of tables produced depends upon the relative type specifed. For all types the following 2 tables are outputted:
\itemize{
\item \code{G_summary1}; table containing counts and proportions of observed major locus genotypes for all relatives of type \code{relative_type},
\item \code{G_summary2}; table containing counts and proportions of observed major locus genotypes for all relatives of type \code{relative_type}, stratified by the major locus genotype of the proband.
}
For relative types \code{"S"}, \code{"O"} and \code{"C"}, 2 additional tables are generated:
\itemize{
\item \code{G_summary3}; table containing counts and proportions of observed major locus genotypes for all relatives of type \code{relative_type}, stratified by sibship size.
\item \code{G_summary4}; table containing counts and proportions of observed major locus genotypes for all relatives of type \code{relative_type}, stratified by the major locus genotype of the proband and sibship size.
}

Summary measures provided do not take into account correlation between individuals from the same family.

Note: If in \code{mimsim.object} non-consenting but eligible families are not stored and/ or complete ascertainment has been specified then \code{include = "Ascertained"} is equivalent to \code{include = "Complete"}.
}
\examples{
# 1. Using Example 1 from mimsim help page.

ped1 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4)

# Summary of G for siblings of probands in ped1
Gsummary1S <- major_locus_G_summary(ped1, relative_type = "S")

# Get slot names
slotNames(Gsummary1S)
# Note: to access a slot type, for example:
# Gsummary1S@G_summary1

# Summary of G for parents of probands in ped1
Gsummary1P <- major_locus_G_summary(ped1, relative_type = "P")

# 2. Using Example 4 from mimsim help page.

ped4 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4,
pConsent=c(0,1), store=TRUE)

# Summary of G for offspring of probands in ped4

Gsummary2aO <- major_locus_G_summary(ped1, relative_type = "O")

# Summary of G for offspring of probands in: 1. pedigree slot and 2. pedigree_unascertained slot

Gsummary2bO <- major_locus_G_summary(ped1, relative_type = "O", include = "Complete")
}
\seealso{
\code{\link[data.table]{data.table}} for information on using data tables.

\code{\link{polygenic_A_summary}} for summary measures of polygenic component \eqn{A} and \code{\link{environment_E_summary}} for summary measures of environmental component \eqn{E}.

\code{\link{simulation_summary}} for estimates of disease risk by relative type.
}

