% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mim_single_ped_binaryFH.R
\name{mim_single_ped_binaryFH}
\alias{mim_single_ped_binaryFH}
\title{Single pedigree mixed inheritance model simulations of disease (1)}
\usage{
mim_single_ped_binaryFH(i, penetrance, maf, h2, r2, eligible = "Genotype",
  relatives = c("P", "S", "GP"), pConsent = c(1, 1), sibship = "Variable",
  mean_sibship = c(2.25, 2.25, 2.25), offspring = TRUE, gene_drop = TRUE,
  store = FALSE, maxit = 1e+06)
}
\arguments{
\item{i}{numeric. Family ID.}

\item{penetrance}{vector. Lifetime penetrance of a bi-allelic major locus \eqn{G}; [\eqn{K_{0}, K_{1}, K_{2}}{K(0), K(1), K(2)}] \eqn{=} [\eqn{p(D = 1 | G = 0), p(D = 1 | G = 1), p(D = 1 | G = 2)}]. Order dependent input.}

\item{maf}{numeric. Risk allele frequency at the major locus, \eqn{G \sim Binom(2, maf)}{G ~ Binom(2, maf)}.}

\item{h2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to a polygenic component, \eqn{A \sim N(0, h2)}{A ~ N(0, h2)}.}

\item{r2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to an environmental component, \eqn{E \sim N(0, r2)}{E ~ N(0, r2)}.}

\item{eligible}{character. Eligibility criteria for ascertainment of a family. 2 options: 1. \code{"Disease"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease, 2. \code{"Genotype"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease \emph{and} carries a risk genotype at the major locus. The default is \code{"Genotype"}.}

\item{relatives}{character vector. Relative types used to calculate family history of disease (see \code{FH}). Can be any combination of \code{"P"} (parents), \code{"S"} (siblings) and \code{"GP"} (grandparents), or \code{NULL}. Default is \code{c("P", "S", "GP")}. Not order dependent.}

\item{pConsent}{vector. The conditional probability that an eligible individual will consent to the study, given family history of disease. \code{pConsent} is a vector of length \eqn{2} and contains the probability for consent given eligible and \eqn{FH = i}, (\eqn{i = 0, 1}). Order dependent input. Default is \code{c(1, 1)}; complete ascertainment.}

\item{sibship}{character. Pedigree structure input. 2 options: 1. \code{"Fixed"} = sibship size is fixed within each generation, 2. \code{"Variable"} = sibship size is variable within and between generations. Default is \code{"Variable"}.}

\item{mean_sibship}{vector. Average sibship size in Generation II, III and IV. Default is \code{c(2.25, 2.25, 2.25)}. Order dependent input.}

\item{offspring}{logical. Pedigree structure input. If \code{TRUE} then the offspring of the proband are generated and outputted families will span \eqn{4} generations. Default is \code{TRUE}.}

\item{gene_drop}{logical. If \code{TRUE} a risk genotype at the major locus is always assigned to 1 individual in Generation I. All other individuals in Generation I are assigned a genotype using the population distribution \eqn{G \sim Bi(2, maf)}{G ~ Bi(2, maf)}. If \code{FALSE} all individuals in Generation I are assigned a major locus genotype using the population distribution. The default is \code{TRUE}.}

\item{store}{logical. If \code{TRUE} eligible but non-consenting families will be stored and outputted. Default is \code{FALSE}.}

\item{maxit}{numeric. Maximum number of iterations. Used in internal functions. Default is \code{1e+06}.}
}
\value{
\code{mim_single_ped_binaryFH} returns either 1. a data frame with 15 columns containing information about the simulated family, or 2. \code{NULL}, meaning there was no individual in Generation III who matched the eligibility criteria specified in \code{eligible}.

\item{Pedigree}{data frame. If there exists \eqn{\geq 1}{\ge 1} individual in Generation III matching the eligibility criteria set in \code{eligible} then a data frame containing pedigree information is outputted. Otherwise output is \code{NULL}.}

If output is a data frame, the information it contains is:
\itemize{
\item Pedigree; the pedigree ID (\code{i}),
\item ID; individual ID,
\item Generation; the generation an individual belongs to (1 - 4);
\item Father; ID of father,
\item Mother; ID of mother,
\item Mate; ID of mate,
\item Sex; gender (1 = Male, 2 = Female),
\item A; polygenic component,
\item G; major locus genotype,
\item E; environmental component,
\item L; liability to disease,
\item D; disease status,
\item child_num; number of offspring,
\item Relationship; relationship to the proband,
\item Ascertained; given there is an eligible individual in the family, if they consent then 1 is assigned (ascertained), else 0 is assigned (unascertained).
}

Because \code{mim_single_ped_binaryFH} can have \code{NULL} output, users are encouraged to use \code{\link{mimsim}} rather than call \code{mim_single_ped_binaryFH} directly. If required, a single family can be generated within \code{mimsim} by setting input \code{n = 1}. See \code{help(mimsim)} for details.
}
\description{
\code{mim_single_ped_binaryFH} is a workhorse function for \code{\link{mimsim}}. Users will not typically call \code{mim_single_ped_binaryFH} directly. It can generate a single, 3 or 4 generation family (pedigree) where \eqn{\geq 1}{\ge 1} individual in Generation III has the disease of interest. Ascertainment of a family can be influenced by disease family history, where family history is a \emph{binary} variable. That is, family history will equal 1 if there exists \eqn{\geq1}{\ge 1} affected relative, and 0 otherwise.
}
\details{
When \code{mimsim} input \code{FH} is set to \code{"binary"}, \code{mim_single_ped_binaryFH} is repeatedly called until the number of families in output equals the required number \code{n}.

Since \code{mim_single_ped_binaryFH} can produce a \code{NULL} output, it is recommended to use \code{\link{mimsim}} rather than call \code{mim_single_ped_binaryFH} directly.
}
\examples{
single_ped <- mim_single_ped_binaryFH(i=1, penetrance = c(0.001, 0.55, 0.55), maf=0.005,
h2 = 0.25, r2 = 0.2)

# Alternatively use mimsim to guarantee a family is outputted.

single_ped <- mimsim(n=1, penetrance = c(0.001, 0.55, 0.55), maf=0.005, h2 = 0.25, r2 = 0.2)

# Pedigree data accessed using:
single_ped@pedigree
}
\seealso{
\code{\link{mimsim}}, which is the recommended function to generate a single pedigree.
\code{\link{mim_single_ped_countFH}} for the generation of a single pedigree using an alternative definition of disease family history.
}

