% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mim_single_ped_countFH.R
\name{mim_single_ped_countFH}
\alias{mim_single_ped_countFH}
\title{Single pedigree mixed inheritance model simulations of disease (2)}
\usage{
mim_single_ped_countFH(i, penetrance, maf, h2, r2, eligible = "Genotype",
  pConsent = c(20, 0, 0, 0), sibship = "Variable", mean_sibship = c(2.25,
  2.25, 2.25), offspring = TRUE, gene_drop = TRUE, store = FALSE,
  maxit = 1e+06)
}
\arguments{
\item{i}{numeric. Family ID.}

\item{penetrance}{vector. Lifetime penetrance of a bi-allelic major locus \eqn{G}; [\eqn{K_{0}, K_{1}, K_{2}}{K(0), K(1), K(2)}] \eqn{=} [\eqn{p(D = 1 | G = 0), p(D = 1 | G = 1), p(D = 1 | G = 2)}]. Order dependent input.}

\item{maf}{numeric. Risk allele frequency at the major locus, \eqn{G \sim Binom(2, maf)}{G ~ Binom(2, maf)}.}

\item{h2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to a polygenic component, \eqn{A \sim N(0, h2)}{A ~ N(0, h2)}.}

\item{r2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to an environmental component, \eqn{E \sim N(0, r2)}{E ~ N(0, r2)}.}

\item{eligible}{character. Eligibility criteria for ascertainment of a family. 2 options: 1. \code{"Disease"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease, 2. \code{"Genotype"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease \emph{and} carries a risk genotype at the major locus. The default is \code{"Genotype"}.}

\item{pConsent}{vector. Defines the conditional probability that an eligible individual will consent to the study, given family history of disease. \code{pConsent} is a vector of length \eqn{4} and contains model parameters for a logistic model of the probability of consent given eligible and the number of affected relatives. A detailed description is given under 'Details'. Order dependent input.}

\item{sibship}{character. Pedigree structure input. 2 options: 1. \code{"Fixed"} = sibship size is fixed within each generation, 2. \code{"Variable"} = sibship size is variable within and between generations. Default is \code{"Variable"}.}

\item{mean_sibship}{vector. Average sibship size in Generation II, III and IV. Default is \code{c(2.25, 2.25, 2.25)}. Order dependent input.}

\item{offspring}{logical. Pedigree structure input. If \code{TRUE} then the offspring of the proband are generated and outputted families will span \eqn{4} generations. Default is \code{TRUE}.}

\item{gene_drop}{logical. If \code{TRUE} a risk genotype at the major locus is always assigned to 1 individual in Generation I. All other individuals in Generation I are assigned a genotype using the population distribution \eqn{G \sim Bi(2, maf)}{G ~ Bi(2, maf)}. If \code{FALSE} all individuals in Generation I are assigned a major locus genotype using the population distribution. The default is \code{TRUE}.}

\item{store}{logical. If \code{TRUE} eligible but non-consenting families will be stored and outputted. Default is \code{FALSE}.}

\item{maxit}{numeric. Maximum number of iterations. Used in internal functions. Default is \code{1e+06}.}
}
\value{
\code{mim_single_ped_countFH} returns either 1. a data frame with 15 columns containing information about the simulated family, or 2. \code{NULL}, meaning there was no individual in Generation III who matched the eligibility criteria specified in \code{eligible}.

\item{Pedigree}{data frame. If \eqn{\geq 1}{\ge 1} individual in Generation III matching the eligibility criteria set in \code{eligible} then a data frame containing pedigree information is outputted. Otherwise output is \code{NULL}.}

If output is a data frame, the information it contains is:
\itemize{
\item Pedigree; the pedigree ID (\code{i}),
\item ID; individual ID,
\item Generation; the generation an individual belongs to (1 - 4);
\item Father; ID of father,
\item Mother; ID of mother,
\item Mate; ID of mate,
\item Sex; gender (1 = Male, 2 = Female),
\item A; polygenic component,
\item G; major locus genotype,
\item E; environmental component,
\item L; liability to disease,
\item D; disease status,
\item child_num; number of offspring,
\item Relationship; relationship to the proband,
\item Ascertained; given there is an eligible individual in the family, if they consent then 1 is assigned (ascertained), else 0 is assigned (unascertained).
}

Because \code{mim_single_ped_countFH} can have \code{NULL} output, users are encouraged to use \code{\link{mimsim}} rather than call \code{mim_single_ped_countFH} directly. If required, a single family can be generated within \code{mimsim} by setting input \code{n = 1}. See \code{help(mimsim)} for details.
}
\description{
\code{mim_single_ped_countFH} is a workhorse function for \code{\link{mimsim}}. Users will not typically call \code{mim_single_ped_countFH} directly. It generates a single, 3 or 4 generation family (pedigree) where \eqn{\geq 1}{\ge 1} individual in Generation III has a disease. Ascertainment of a family can be influenced by disease family history, where family history is a \emph{count} of the number of siblings, parents and grandparents who are affected.
}
\details{
\code{mim_single_ped_countFH} aims to output a single family which meets the eligibility criteria specified in \code{eligible}. If the simulated family does not meet this criteria then output is \code{NULL}.

When \code{mimsim} input \code{FH} is set to \code{"count"}, \code{mim_single_ped_countFH} is repeatedly called until the number of families in output equals the required number \code{n}.

Since \code{mim_single_ped_countFH} can produce a \code{NULL} output, it is recommended to use \code{\link{mimsim}} rather than call \code{mim_single_ped_countFH} directly.

Input variable \code{pConsent = c(a, b, c, d)}; a vector of length 4 used to define the probability of consent of an individual given: 1. they are eligible and 2. the disease status of \code{relatives}. The probability of consent given eligible and the disease status of \code{relative}, is calculated using the following logistic model:
\deqn{logit\big(p(Consent | Eligible, D(relatives))\big) = a + b\sum_{i=1}^{2}D_{"P", i} + c\sum_{j=1}^{S_{e} - 1}D_{"S", j} + d\sum_{k=1}^{4}D_{"GP", k} ,}{logit(p(Consent | Eligible, D(relatives))) = a + b\sumD("P", i) + c\sumD("S", j), d\sumD("GP", k) ,}
 where:
\itemize{
\item \eqn{D_{"P", i}}{D("P", i)} is a binary variable, which equals \eqn{1} if the \eqn{i^{th}}{i-th} parent of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{i = 1, 2},
\item \eqn{D_{"S", j}}{D("S", j)} is a binary variable, which equals \eqn{1} if the \eqn{j^{th}}{j-th} sibling of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{j = 1, 2, ..., S_{e} - 1}{j = 1, 2, ..., S(e) - 1}, and where \eqn{S_{e}}{S(e)} is sibship size of the sibship containing the eligible individual under investigation,
\item \eqn{D_{"GP", k}}{D("GP", k)} is a binary variable, which equals \eqn{1} if the \eqn{k^{th}}{k-th} grandparent of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{k = 1, 2, 3, 4},
\item \code{a} defines the probability of consent for eligible individuals with no family history of disease,
\item \code{b} defines the increase in \eqn{logit} consent probability for each additional parent (\code{"P"}) affected with the disease,
\item \code{c} defines the increase in \eqn{logit} consent probability for each additional sibling (\code{"S"}) affected with the disease, and,
\item \code{d} defines the increase in \eqn{logit} consent probability for each additional grandparent (\code{"GP"}) affected with the disease.
}
The default value of \code{pConsent = c(20, 0, 0, 0)} sets the probability of consent for every eligible individual equal to 1.

The user can explore the consent probabilities used by \code{mim_single_ped_countFH} with the function \code{\link{probability_consent_count}}. This function calculates the consent probability for a user-specified \code{pConsent}, \code{n_D_P} \eqn{= \sum_{i=1}^{2}D_{"P", i}}{= \sumD("P", i)}, \code{n_D_S} \eqn{= \sum_{j=1}^{S_{e} - 1}D_{"S", j}}{= \sumD("S", j)} and \code{n_D_GP} \eqn{= \sum_{k=1}^{4}D_{"GP", k}}{= \sumD("GP", k)}.
}
\examples{
single_ped <- mim_single_ped_countFH(i=1, penetrance = c(0.001, 0.55, 0.55), maf=0.005,
h2 = 0.25, r2 = 0.2)

# Alternatively use mimsim to guarantee a family is outputted.

single_ped <- mimsim(n=1, penetrance = c(0.001, 0.55, 0.55), maf=0.005, h2 = 0.25, r2 = 0.2,
FH = "count", pConsent = c(20, 0, 0, 0))

# Pedigree data accessed using:
single_ped@pedigree
}
\seealso{
\code{\link{mimsim}}, which is the recommended function to generate a single pedigree.

\code{\link{mim_single_ped_binaryFH}} for the generation of a single pedigree using an alternative definition of disease family history.

\code{\link{probability_consent_count}} allows users to explore the probability of consent used in \code{mim_single_ped_countFH}.
}

