% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mimsim.R
\name{mimsim}
\alias{mimsim}
\title{Mixed inheritance model simulation of disease within families}
\usage{
mimsim(n, penetrance, maf, h2, r2, eligible = "Genotype", FH = "binary",
  relatives = c("P", "S", "GP"), pConsent = c(1, 1), sibship = "Variable",
  mean_sibship = c(2.25, 2.25, 2.25), offspring = TRUE, p = 1,
  store = FALSE, maxit = 1e+06)
}
\arguments{
\item{n}{numeric. Number of families to be generated.}

\item{penetrance}{vector. Lifetime penetrance of a bi-allelic major locus \eqn{G}; [\eqn{K_{0}, K_{1}, K_{2}}{K(0), K(1), K(2)}] \eqn{=} [\eqn{p(D = 1 | G = 0), p(D = 1 | G = 1), p(D = 1 | G = 2)}]. Order dependent input.}

\item{maf}{numeric. Risk allele frequency at the major locus, \eqn{G \sim Binom(2, maf)}{G ~ Binom(2, maf)}.}

\item{h2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to a polygenic component, \eqn{A \sim N(0, h2)}{A ~ N(0, h2)}.}

\item{r2}{numeric. Proportion of variability in liability to disease, \eqn{L}, attributable to an environmental component, \eqn{E \sim N(0, r2)}{E ~ N(0, r2)}.}

\item{eligible}{character. Eligibility criteria for ascertainment of a family. 2 options: 1. \code{"Disease"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease, 2. \code{"Genotype"} = family is eligible if there exists \eqn{\geq 1}{\ge 1} individual in Generation III who has the disease \emph{and} carries a risk genotype at the major locus. The default is \code{"Genotype"}.}

\item{FH}{character. Family history of disease definition. 2 options: 1. \code{"binary"} = family history variable is binary and indicates if any relative has the disease, 2. \code{"count"} = family history variable counts the number of affected relatives. Relatives included in the definition of family history are specified using input variable \code{relatives}.}

\item{relatives}{character vector. Relative types used in to calculate family history of disease (see \code{FH}). Can be any combination of \code{"P"} (parents), \code{"S"} (siblings) and \code{"GP"} (grandparents), or \code{NULL}. Default is \code{c("P", "S", "GP")}. Not order dependent.}

\item{pConsent}{vector. Defines the conditional probability that an eligible individual will consent to the study, given family history of disease. 2 options depending on \code{FH} input. 1. If \code{FH = "binary"} (default), \code{pConsent} is a vector of length \eqn{2} and contains the probability for consent given eligible and \eqn{FH = i}, (\eqn{i = 0, 1}). 2. If \code{FH = "count"}, \code{pConsent} is a vector of length \eqn{4} and contains model parameters for a logistic model of the probability of consent given eligible and the number of affected relatives. A detailed description is given under 'Details'. Order dependent input.}

\item{sibship}{character. Pedigree structure input. 2 options: 1. \code{"Fixed"} = sibship size is fixed within each generation, 2. \code{"Variable"} = sibship size is variable within and between generations. Default is \code{"Variable"}.}

\item{mean_sibship}{vector. Average sibship size in Generation II, III and IV. Default is \code{c(2.25, 2.25, 2.25)}. Order dependent input.}

\item{offspring}{logical. Pedigree structure input. If \code{TRUE} then the offspring of the proband are generated and outputted families will span \eqn{4} generations. Default is \code{TRUE}.}

\item{p}{numeric. Probability that gene dropping is used. Default is 1; gene dropping is always used.}

\item{store}{logical. If \code{TRUE} eligible but non-consenting families will be stored and outputted. Default is \code{FALSE}.}

\item{maxit}{numeric. Simulation stops when: a. \code{n} families have been generated or b. iteration number\eqn{=} \code{maxit}. Default is \code{1e+06}. Substantially reducing this number may lead to \eqn{<} \code{n} families in \code{mimsim} output.}
}
\value{
\code{mimsim} returns an S4 class object. To view a slot type: \code{mimsim.object@slot_name}.

Slots are:
\item{pedigree}{data frame. Pedigree data. Has \eqn{14} columns which includes variables such as individual ID (\code{ID}), ID of parents (\code{Father} and \code{Mother}), liability to disease (\code{L}) and disease status (\code{D}).}
\item{proband_info}{data frame. Information about probands. Has \eqn{21} columns which includes variables such as a binary variable indicating whether any 1st degree relatives (defined as parents and siblings) are affected (\code{FH1}), total number of relatives of type \code{R} (\code{n_R}) and total number affected relatives of type \code{R} (\code{n_D_R}), for \code{R} = \code{P} (parents), \code{S} (siblings), \code{O} (offspring), \code{GP} (grandparents), \code{Av} (aunts and uncles by blood), \code{MAv} (aunts and uncles by marriage), \code{C} (cousins) and \code{M} (mate).}
\item{pedigree_unascertained}{data frame. Pedigree data for eligible but non-consenting families. If \code{store = FALSE} (default) or \code{store = TRUE} but no such families exist then \code{NA} is returned. Else, a data frame of same structure as \code{pedigree} is returned.}
\item{proband_info_unascertained}{data frame. Proband information for eligible but non-consenting families. If \code{store = FALSE} (default) or \code{store = TRUE} but no such families exist then \code{NA} is returned. Else, a data frame of same structure as \code{proband_info} is returned.}
\item{D_threshold}{numeric. Disease threshold, \eqn{T}; \eqn{T = \sqrt{h2 + r2}\Phi^{-1}\big(1 - K_{0}\big)}{T = \sqrt(h2 + r2)\Phi^(-1)[1 - K(0)]}, where \eqn{K_{0} = p(D = 1 | G = 0)}{K(0) = p(D = 1 | G = 0)} and is the first element in vector \code{penetrance}.}
\item{beta}{matrix. Mean shift in the distribution of liability to disease, \eqn{L}, when [\eqn{G = 0, G = 1, G = 2}]. \eqn{G=0} is the reference category; \eqn{\beta_{0} = 0}{\beta(0) = 0}. \eqn{\beta_{i} = T - \sqrt{h2 + r2}\Phi^{-1}\big(1 - K_{i}\big)}{\beta(i) = T - sqrt(h2 + r2)\Phi^(-1)[1 - K(i)]}, where \eqn{K_{i} = p(D = 1 | G = i)}{K(i) = p(D = 1 | G = i)} for \eqn{i = 1, 2} are specified within \code{penetrance}.}
\item{G_MOI}{character. Mode of inheritance for the major locus \eqn{G}. E.g. if \code{penetrance = c(0.1, 0.6, 0.6)}, then \code{G_MOI = "Dominant"}.}
\item{L_given_G_0}{matrix. Distribution parameters for conditional liability to disease, \eqn{L |G = 0}; \eqn{L | \{G = 0\} \sim N(0, h2 + r2)}{L | (G = 0) ~ N(0, h2 + r2)}.}
\item{L_given_G_1}{matrix. Distribution parameters for conditional liability to disease, \eqn{L | G = 1}; \eqn{L | \{G = 1\} \sim N(\beta_{1}, h2 + r2)}{L | (G = 1) ~ N(\beta(1), h2 + r2)}.}
\item{L_given_G_2}{matrix. Distribution parameters for conditional liability to disease, \eqn{L | G = 2}; \eqn{L | \{G = 2\} \sim N(\beta_{2}, h2 + r2)}{L | (G = 2) ~ N(\beta(2), h2 + r2)}.}
\item{L}{matrix. Distribution parameters for liability to disease, \eqn{L}. \code{L} is a mixture of the conditional distributions specified in \code{L_given_G_0}, \code{L_given_G_1} and \code{L_given_G_2}.}
\item{prevalence}{numeric. Lifetime population prevalence of disease.}
\item{...}{Input arguments. E.g. To view input variable \code{eligible} type: \code{mimsim_object_name@eligible}.
Note: logical variables are changed such that \code{TRUE} = \code{Yes} and \code{FALSE} = \code{No}.}
}
\description{
\code{mimsim} uses a mixed inheritance model to simulate disease within \eqn{3} or \eqn{4} generation families. Families are generated forwards-in-time. Outputted families contain a \emph{proband}; an eligible and consenting individual in Generation III. The probability that an eligible individual consents can be a function of the family history of disease, allowing the user to generate family data when there is ascertainment bias.
}
\details{
\code{mimsim} inputs can broadly be categorised into 3 classes:

\bold{1.} Disease model inputs: \code{penetrance}, \code{maf}, \code{h2} and \code{r2}.
The disease model is a mixed inheritance model (MIM); a liability threshold approach where covariates are linked to disease status, \eqn{D}, through a continuous variable called liability to disease, \eqn{L}. If \eqn{L > T} then \eqn{D = 1}, else if \eqn{L \leq T}{L \le T} then \eqn{D = 0}, where \eqn{T} is the disease threshold. Within \code{mimsim} \eqn{L} is a mixture of the following normal distributions: \deqn{L | \{G = g\} = \beta_{g} + A + E ~ \sim N(0, h2 + r2)~,}{L | (G = g) = \beta(g) + A + E  ~ N(0, h2 + r2) ,} for \eqn{g = 0, 1, 2}. Where polygenic component \eqn{A \sim N(0, h2)}{A ~ N(0, h2)}, environmental component \eqn{E \sim N(0, r2)}{E ~ N(0, r2)} and \eqn{\beta_{0} = 0}{\beta(0) = 0}.
Disease model inputs are used to: \itemize{
\item calculate \eqn{T}, \eqn{\beta_{1}}{\beta(1)} and \eqn{\beta_{2}}{\beta(2)}, and,
\item simulate \eqn{A}, \eqn{E} and \eqn{G} for each family member, and thus calculate \eqn{L} and \eqn{D}.
}

\bold{2.} Family structure inputs: \code{sibship}, \code{mean_sibship} and \code{offspring}.
Together \code{sibship} and \code{mean_sibship} specify sibship sizes across a family. \code{offspring} deterimines whether a family will span \eqn{3} or \eqn{4} generations.
If \code{sibship = "Fixed"} then sibship size, \eqn{S}, is fixed \emph{within} each generation. \eqn{S} may still vary \emph{between} generations, for example, if \code{mean_sibship = c(3, 2, 1)} then \eqn{S_{II} = 3}{S(II) = 3} for all sibships in Generation II, \eqn{S_{III} = 2}{S(III) = 2} for all sibships in Generation III, and \eqn{S_{IV} = 1}{S(IV) = 1} for all sibships in Generation IV. When \code{sibship = "Fixed"} and a non-integer value is specified within \code{mean_sibship} then the ceiling function is used, i.e. \code{mean_sibship = c(2.25, 2.25, 2.25)} will become \code{mean_sibship = c(3, 3, 3)}.
When \code{sibship = "Variable"}, the size of each sibship within a family is a random draw from a Poisson distribution, with a common mean within a generation specified by \code{mean_sibship}. For example, if \code{sibship = "Variable"} and \code{mean_sibship = c(3, 2, 1)} then: \eqn{S_{II} \sim Pois(3)}{S(II) ~ Pois(3)}, \eqn{S_{III} \sim Pois(2)}{S(III) ~ Pois(2)} and \eqn{S_{IV} \sim Pois(1)}{S(IV) ~ Pois(1)}.

\bold{3.} Ascertainment inputs: \code{eligible}, \code{FH}, \code{relatives} and \code{pConsent}.
These inputs set criteria for which families should be outputted. \code{mimsim} mimics some aspects of ascertainment for family studies investigating lifetime risk of disease. Users set eligibility criteria, applied in Generation III, as either: 1. \code{eligible = "Disease"}; an eligible individual has \eqn{D = 1}, or 2. \code{eligible = "Genotype"}; an eligible individual has \eqn{\{D = 1, G = risk genotype\}}{{D = 1, G = risk genotype}} (a penetrance study). Not all eligible individuals will \emph{consent} to the study, and the decision to consent may be influenced by family history of disease. That is, there may be ascertainment bias. This can be specified using \code{FH}, \code{relatives} and \code{pConsent}. \code{FH} specifies whether family history is defined as a binary variable (\code{FH = "binary"}) or a count variable (\code{FH = "count"}). \code{relatives} specifies which relatives are used in the definition of family history. This can be: 1. any combination of parents (\code{"P"}), siblings (\code{"S"}) and/ or grandparents (\code{"GP"}), or, 2. \code{NULL}.
\code{pConsent} is a vector used to define the probability of consent. It's length and meaning differ depending on \code{FH}. When:

\emph{(i)}  \code{FH = "binary"} then \code{pConsent = c(a, b)}; a vector of length 2 where \code{a} is the probability of consent given eligible and no family history of disease (\eqn{a = p(Consent | Eligible, FH = 0)}) and \code{b} is the probability of consent given eligible and there is a family history of disease (\eqn{a = p(Consent | Eligible, FH = 1)}).

\emph{(ii)} \code{FH = "count"} then \code{pConsent = c(a, b, c, d)}; a vector of length 4 used to define the probability of consent of an individual given: 1. they are eligible and 2. the disease status of \code{relatives}. The probability of consent given eligible and the disease status of \code{relative}, is calculated using the following logistic model:
\deqn{logit\big(p(Consent | Eligible, D(relatives))\big) = a + b\sum_{i=1}^{2}D_{"P", i} + c\sum_{j=1}^{S_{e} - 1}D_{"S", j} + d\sum_{k=1}^{4}D_{"GP", k} ,}{logit(p(Consent | Eligible, D(relatives))) = a + b\sumD("P", i) + c\sumD("S", j), d\sumD("GP", k) ,}
 where:
\itemize{
\item \eqn{D_{"P", i}}{D("P", i)} is a binary variable, which equals \eqn{1} if the \eqn{i^{th}}{i-th} parent of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{i = 1, 2},
\item \eqn{D_{"S", j}}{D("S", j)} is a binary variable, which equals \eqn{1} if the \eqn{j^{th}}{j-th} sibling of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{j = 1, 2, ..., S_{e} - 1}{j = 1, 2, ..., S(e) - 1}, and where \eqn{S_{e}}{S(e)} is sibship size of the sibship containing the eligible individual under investigation,
\item \eqn{D_{"GP", k}}{D("GP", k)} is a binary variable, which equals \eqn{1} if the \eqn{k^{th}}{k-th} grandparent of the proband is affected with the disease, and \eqn{0} otherwise; for \eqn{k = 1, 2, 3, 4},
\item \code{a} defines the probability of consent for eligible individuals with no family history of disease,
\item \code{b} defines the increase in \eqn{logit} consent probability for each additional parent (\code{"P"}) affected with the disease,
\item \code{c} defines the increase in \eqn{logit} consent probability for each additional sibling (\code{"S"}) affected with the disease, and,
\item \code{d} defines the increase in \eqn{logit} consent probability for each additional grandparent (\code{"GP"}) affected with the disease.
}
If user is not confident with using \code{FH = "count"} due to uncertainty over what consent probabilities are being used, please see \code{\link{probability_consent_count}}. This function allows the user to explore consent probabilities for user-specified \code{pConsent} and \code{n_D_P} \eqn{= \sum_{i=1}^{2}D_{"P", i}}{= \sumD("P", i)}, \code{n_D_S} \eqn{= \sum_{j=1}^{S_{e} - 1}D_{"S", j}}{= \sumD("S", j)} and \code{n_D_GP} \eqn{= \sum_{k=1}^{4}D_{"GP", k}}{= \sumD("GP", k)}.

Note: If \code{relatives = NULL} then everyone is assumed to have no family history of disease. In this instance to save computational time it is recommended to use complete ascertainment. The simplest way to do this is to use the default settings: \code{FH = "binary"} and \code{pConsent = c(1, 1)}.
}
\section{Additional information}{

1. Error given when inputs \code{penetrance}, \code{maf}, \code{h2} and \code{r2} result in \eqn{var(L) > 1}.

2. When \code{maf = 0} regardless of specification in inputs, \code{eligible = "Disease"} and \code{p = 0}.
}
\examples{
# 1. Penetrance study: eligible = "Genotype" (default so not typed).
#     Complete ascertainment: no ascertainment bias.
#     100 families.
#     Dominant major locus. prevalence = 0.01688.
#     Variable sibship size (sibship = "Variable") with average sibship size = 2.25
#     pedigree-wide.

ped1 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4)


# 2. Family study of lifetime disease risk in families. eligible = "Disease".
#     No major locus carriers: maf = 0.
#     Note 1: when maf = 0, regardless of p, no gene dropping is used.
#     mean_sibship=c(2.2, 2, 1): average sibship size differ across generations.
#     Note 2: ped2a has penetrance = c(0.01, 0.2, 0.2) and ped2b has
#     penetrance = c(0.01, 0.01, 0.01). This does not matter as long as 1st element
#     in penetrance = p(D = 1 | G = 0) is the same.

ped2a <- mimsim(n=100, penetrance=c(0.01, 0.2, 0.2), maf=0, h2=0.7, r2=0.2 ,
eligible="Disease", mean_sibship=c(2.2,2,1))

ped2b <- mimsim(n=100, penetrance=c(0.01, 0.01, 0.01), maf=0, h2=0.7, r2=0.2 ,
eligible="Disease", mean_sibship=c(2.2,2,1))


# 3. Penetrance study: eligible = "Genotype".
#     Sibship = "Fixed", mean_sibship = c(2, 2, 2).
#     Recessive major locus (see penetrance).
ped3 <- mimsim(n = 50, penetrance = c(0.005, 0.005, 0.8), maf=0.001, h2=0.5, r2=0.1,
sibship="Fixed", mean_sibship=c(2,2,2))


# 4. Example 1 with ascertainment bias:
#    FH = "binary" (default so not typed),
#    relatives = c("P", "S", "GP") (default so not typed).
#    pConsent = c(0.8, 1) : eligible individuals with no family history of
#    disease consent 80\% of the time, eligible individuals with a family
#    history of disease always consent.
#    store = TRUE (stored eligible but non-consenting families).

ped4 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4,
pConsent=c(0.8,1), store=TRUE)


# 5. Penetrance study: eligible = "Genotype".
#    with ascertainment bias:
#    FH = "count",
#    pConsent = c(0.5, 1, 1, 0.5), this means, for example:
# - p(Consent | Eligible, no FH) = 0.62,
# - p(Consent | Eligible, {1 affected parent OR 1 affected sibling}) = 0.82,
# - p(Consent | Eligible, {2 affected parent OR 1 affected parent, 1 affected sibling}) = 0.92,
# - p(Consent | Eligible, 1 affected grandparent) = 0.73,
# - p(Consent | Eligible, 1 affected parents, 1 affected sibling, 1 affected grandparent) = 0.95.
#     store = TRUE (stored eligible but non-consenting families).
#     Dominant major locus.

ped5 <- mimsim(n = 100, penetrance = c(0.001, 0.55, 0.55), maf=0.005, h2 = 0.25, r2 = 0.2,
FH="count", pConsent=c(0.5, 1, 1, 0.5), store=TRUE)


# 6. Example of combination of polygenic component, environmental component
#     and major locus giving var(liability) > 1
\dontrun{pederror <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.05, h2=0.3, r2=0.4)}
}
\seealso{
\code{\link{polygenic_A_summary}}, \code{\link{major_locus_G_summary}} and \code{\link{environment_E_summary}} produce summary measures for the polygenic component \eqn{A}, the major locus \eqn{G} and the environmental component \eqn{E} respectively.

\code{simulation_summary} for estimates of disease risk by relative type.

\code{\link{probability_consent_count}} allows users to explore the probability of consent used by \code{mimsim} when \code{FH = "count"}.
}

