% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s4c_Cousin_G_summary.R
\docType{class}
\name{s4c_Cousin_G_summary}
\alias{s4c_Cousin_G_summary}
\alias{s4c_Cousin_G_summary-class}
\title{An S4 class for \code{major_locus_G_summary} output: cousins}
\description{
This is an internal S4 class within \code{\link{major_locus_G_summary}}.
}
\section{Slots}{

\describe{
\item{\code{Description}}{character. Brief overview of output.}

\item{\code{G_summary1}}{data.table. Counts and proportions of observed major locus genotypes for cousins.}

\item{\code{G_summary2}}{data.table. Counts and proportions of observed major locus genotypes for cousins, stratified by the major locus genotype of the proband.}

\item{\code{G_summary3}}{data.table. Counts and proportions of observed major locus genotypes for cousins, stratified by sibship size.}

\item{\code{G_summary4}}{data.table. Counts and proportions of observed major locus genotypes for cousins, stratified by the major locus genotype of the proband and sibship size.}
}}

