% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s4c_marginal_output.R
\docType{class}
\name{s4c_marginal_output}
\alias{s4c_marginal_output}
\alias{s4c_marginal_output-class}
\title{An S4 class representing disease model parameters and the distribution of liability to disease}
\description{
This is an internal S4 class within \code{\link{mimsim}}.
}
\section{Slots}{

\describe{
\item{\code{model_parameters}}{matrix. Contains the following disease model parameters: \code{D_threshold} (\eqn{t}), \code{beta_0} (\eqn{beta_0}{beta(0)}), \code{beta_1} (\eqn{beta_1}{beta(1)}) and \code{beta_2} (\eqn{beta_2}{beta(2)}).}

\item{\code{marginal_distn}}{matrix. Contains the \code{Mean} and \code{Variance} of liability to disease \eqn{L}.}
}}

