% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s4c_mimsim_output.R
\docType{class}
\name{s4c_mimsim_output}
\alias{s4c_mimsim_output}
\alias{s4c_mimsim_output-class}
\title{An S4 class for \code{mimsim} output}
\description{
This is an internal S4 class within \code{\link{mimsim}}, used to store \code{mimsim} output. Inherits slots from S4 class \code{input_normal}.
}
\section{Slots}{

\describe{
\item{\code{pedigree}}{data.frame. Contains ascertained families.}

\item{\code{proband_info}}{data.frame. Contains additional information for the proband.}

\item{\code{pedigree_unascertained}}{data.frame. Contains eligible but unascertained probands.}

\item{\code{proband_info_unascertained}}{data.frame. Contains additional information for the unascertained but eligible probands.}
}}

