% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s4c_Parent_G_summary.R
\docType{class}
\name{s4c_Parent_G_summary}
\alias{s4c_Parent_G_summary}
\alias{s4c_Parent_G_summary-class}
\title{An S4 class representing output from major_locus_G_summary for the parents of the proband}
\description{
This is an internal S4 class within \code{\link{major_locus_G_summary}}.
}
\section{Slots}{

\describe{
\item{\code{Description}}{character. Brief overview of output.}

\item{\code{G_summary1}}{data.table. Counts and proportions of observed major locus genotypes for parents.}

\item{\code{G_summary2}}{data.table. Counts and proportions of observed major locus genotypes for parents, stratified by the major locus genotype of the proband.}
}}

