% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s4c_polygenic_summary.R
\docType{class}
\name{s4c_polygenic_summary}
\alias{s4c_polygenic_summary}
\alias{s4c_polygenic_summary-class}
\title{An S4 class for \code{polygenic_A_summary} output}
\description{
This is an internal S4 class within \code{\link{polygenic_A_summary}}.
}
\section{Slots}{

\describe{
\item{\code{A}}{data.table. Summary measures for the distribution of polygenic component \eqn{A} by relative type.}

\item{\code{A_G0}}{data.table. Summary measures for the distribution of polygenic component \eqn{A} for individuals with \eqn{G = 0}, by relative type. That is, the distribution of \eqn{A | G = 0}, by relative type.}

\item{\code{A_G1}}{data.table. Summary measures for the distribution of polygenic component \eqn{A} for individuals with \eqn{G = 1}, by relative type. That is, the distribution of \eqn{A | G = 1}, by relative type.}

\item{\code{A_G2}}{data.table. Summary measures for the distribution of polygenic component \eqn{A} for individuals with \eqn{G = 2}, by relative type. That is, the distribution of \eqn{A | G = 2}, by relative type.}

\item{\code{A_Carrier}}{data.table. Summary measures for the distribution of polygenic component \eqn{A} for individuals carrying a risk genotype at the major locus, by relative type. That is, the distribution of \eqn{A | G \geq 1}{A | G \ge 1}, by relative type.}
}}

