% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_summary.R
\name{simulation_summary}
\alias{simulation_summary}
\title{Summary of mimsim simulation output}
\usage{
simulation_summary(mimsim.object, include = "Ascertained")
}
\arguments{
\item{mimsim.object}{S4 class object resulting from a call to \code{mimsim}.}

\item{include}{character. 2 options: 1. \code{"Ascertained"} = include families containing an eligible and consenting individual (ascertained families). These families are stored in \code{mimsim.object@pedigree}. 2. \code{"Complete"} = include families containing an eligible individual. This combines families stored in \code{mimsim.object@pedigree} (consenting, and so ascertained, families) and \code{mimsim.object@pedigree_unascertained} (non-consenting, and so unascertained, families). Default is \code{include = "Ascertained"}.}
}
\value{
\code{simulation_summary} returns an S4 class object. To view a slot type: \code{mimsim.summary.object@slot_name}.

Slots are:
\item{RRR}{data table. For all family members, estimated probability of disease (column \code{K}) and recurrence risk ratio (column \code{lambda}) by relative type.}
\item{RRR_G0}{data table. For family members with major locus genotype \eqn{G = 0}, estimated probability of disease (column \code{K_G0}) and recurrence risk ratio (column \code{lambda_G0}) by relative type.}
\item{RRR_G1}{data table. For family members with major locus genotype \eqn{G = 1}, estimated probability of disease (column \code{K_G1}) and recurrence risk ratio (column \code{lambda_G1}) by relative type.}
\item{RRR_G2}{data table. For family members with major locus genotype \eqn{G = 2}, estimated probability of disease (column \code{K_G2}) and recurrence risk ratio (column \code{lambda_G2}) by relative type.}
\item{RRR_carrier}{data table. For family members with major locus genotype \eqn{G \geq 1}{G \ge 1}, estimated probability of disease (column \code{K_carrier}) and recurrence risk ratio (column \code{lambda_carrier}) by relative type.}
\item{Count_summary}{data table. For each relative type, \code{Count_summary} provides a count of the number of families containing \eqn{\geq 1}{\ge 1} relative of type \eqn{R} (column \code{N_peds}), and the number of families containing \eqn{\geq 1}{\ge 1} affected relative of type \eqn{R} (column \code{N_peds_ge1_D}).}

See 'Details' section for relative types included.
}
\description{
\code{simulation_summary} produces estimates of the probability of disease and recurrence risk ratio for relatives of a proband using output from \code{mimsim}. Estimates of disease risk and recurrence risk ratios stratified by major locus genotype, \eqn{G}, are also given.
}
\details{
Note: For all summaries, relative types are \eqn{R} = \code{"P"} (parents), \code{"S"} (siblings), \code{"O"} (offspring), \code{"GP"} (grandparents), \code{"Av"} (aunts and uncles by blood), \code{"C"} (cousins), \code{"MAv"} (aunts and unclues by marriage) and \code{"M"} (mates).

\code{simulation_summary} produces simple estimates of the probability of disease for relatives of type \eqn{R}, \eqn{K_{R}}{K(R)}, for families generated using \code{\link{mimsim}}. Estimates of \eqn{K_{R}}{K(R)} are calculated using:
\deqn{E[K_{R}] = E[p(D_{R} = 1 | ascertained)] = \frac{\sum_{i = 1}^{N_{R}}d_{R, i}}{N_{R}} ~ ,}{E[K(R)] = E[p(D(R) = 1 | ascertained)] = \sum d(R, i)/N(R) ,}
where:
\itemize{
\item \eqn{D_{R}}{D(R)} is the disease status variable for a relative of type \eqn{R},
\item \eqn{d_{R, i}}{d(R, i)} is the observed disease status for the \eqn{i^{th}}{i-th} relative of type \eqn{R} in \code{mimsim.object}, and,
\item \eqn{N_{R}}{N(R)} is the total number of relatives of type \eqn{R} in \code{mimsim.object}.
}
Estimates of the recurrence risk ratio for relatives of type \eqn{R}, \eqn{\lambda_{R}}{\lambda(R)} are calculated using:
\deqn{E[\lambda_{R}] = \frac{E[K_{R}]}{K} ~ ,}{E[\lambda(R)] = E[K(R)]/K ,}
where \eqn{K = p(D = 1)} is the population prevalence of disease.

Estimates of disease risk to relatives of a proband, stratified by major locus genotype (\eqn{G}) are calculated using:
\deqn{E[K_{R | G_{R} = g}] = E[p(D_{R} = 1 | G_{R} = g, ascertained)] = \frac{\sum_{i = 1}^{N_{R | G_{R} = g}} d_{R | G_{R} = g, i}}{N_{R | G_{R} = g}} ~ ,}{E[K(R | G(R) = g)] = E[p(D(R) = 1 | G(R) = g, ascertained)] = \sum d(R | G(R) = g, i)/N(R | G(R) = g) ,}
for \eqn{g = 0, 1, 2}, and where:
\itemize{
\item \eqn{D_{R}}{D(R)} is the disease status variable for a relative of type \eqn{R},
\item \eqn{G_{R}}{G(R)} is the major locus genotype variable for a relative of type \eqn{R},
\item \eqn{d_{R | G_{R} = g, i}}{d(R | G(R) = g, i)} is the observed disease status for the \eqn{i^{th}}{i-th} relative of type \eqn{R} who has a major locus genotype equal to \eqn{g} in \code{mimsim.object}, and,
\item \eqn{N_{R | G_{R} = g}}{N(R | G(R) = g)} is the total number fo relatives of type \eqn{R} who have a major locus genotype equal to \eqn{g} in \code{mimsim.object}.
}
Estimates of recurrence risk ratio for relatives of a proband, stratified by major locus genotype (\eqn{G}) are calculated using:
\deqn{E[\lambda_{R | G_{R} = g}] = \frac{E[K_{R | G_{R} = g}]}{K_{g}} ~ ,}{E[\lambda(R | G(R) = g)] = E[K(R | G(R) = g)]/K(g) ,}
for \eqn{g = 0, 1, 2}, and where \eqn{K_{g} = p(D = 1 | G = g)}{K(g) = p(D = 1 | G = g)}.

Summary measures provided do not take into account correlation between individuals from the same family.

Note: If in \code{mimsim.object} non-consenting but eligible families are not stored and/ or complete ascertainment has been specified then \code{include = "Ascertained"} is equivalent to \code{include = "Complete"}.
}
\examples{
# 1. Using Example 1 from mimsim help page.

ped1 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4)

summary1 <- simulation_summary(ped1)

# Get slot names
slotNames(summary1)
# Note: to access a slot type, for example:
# summary1@RRR

# 2. Using Example 4 from mimsim help page.

ped4 <- mimsim(n=100, penetrance=c(0.01, 0.7, 0.7), maf=0.005, h2=0.3, r2=0.4,
pConsent=c(0,1), store=TRUE)

# Estimates of disease risk and recurrence risk ratio for ascertained families

summary2a <- simulation_summary(ped4)

# Estimates of disease risk and recurrence risk ratio for ascertained families and
# eligible but non-consenting families

summary2b <- simulation_summary(ped4, include = "Complete")
}
\seealso{
\code{\link[data.table]{data.table}} for information on using data tables.

\code{\link{polygenic_A_summary}}, \code{\link{major_locus_G_summary}} and \code{\link{environment_E_summary}} produce summary measures for the polygenic component \eqn{A}, the major locus \eqn{G} and the environmental component \eqn{E} respectively.
}

